[IGOWEB-990] Change number of reads from int to long (#28) (#30)

+++ src/main/java/org/mskcc/kickoff/validator/RequestValidator.java
@@ -240,7 +240,7 @@ public class RequestValidator {
     private void validateReadCounts(Request request) {
         // For each sample in readsForSample if the number of reads is less than 1M, print out a warning
         for (Sample sample : getSamplesWithSampleQc(request)) {
-            int numberOfReads = sample.getNumberOfReads();
+            long numberOfReads = sample.getNumberOfReads();
             if (numberOfReads < Constants.MINIMUM_NUMBER_OF_READS) {
                 // Print AND add to readme file
                 String extraReadmeInfo = String.format("\n[WARNING] sample %s has less than 1M total reads: %d\n", sample, numberOfReads);
